added main

+++ general/trinity_new.m
@@ -1,10 +1,18 @@
 function trinity_new(fn)
-% TRINITY_NEW  Start new Trinity project
+% TRINITY_NEW  Start a new Trinity project
+%   TRINITY_NEW(PROJECT_NAME) makes a new script file to use Trinity.
+
+% (c)2013- Joachim Vandekerckhove. See license.txt for licensing information.
 
 if ~nargin
     fn = 'main';
 end
 
 newname = sprintf('%s_trinity.m', fn);
-copyfile(which('trinity_generic_script'), newname)
+if exist(newname, 'file')
+    warning('trinity:trinity_new:existingFile', ...
+        'A file with that name already exists!')
+else
+    copyfile(which('trinity_generic_script'), newname)
+end
 edit(newname)
